Modificación del gate y policy (se agregó que solo el usuario puede borrar unicamente las mascotas que él publicó)

+++ app/Http/Controllers/MascotaController.php
@@ -123,6 +123,7 @@ class MascotaController extends Controller implements HasMiddleware
      */
     public function destroy(Mascota $mascota)
     {
+        Gate::authorize('delete', $mascota);
         $mascota->delete();
         return redirect()->route('mascota.index');
     }
